uploadtophotobanks/requirements.txt
colorlog>=6.7.0
tqdm>=4.64.0
paramiko>=3.0.0
Pillow>=9.0.0